feat: move new jetbrains features to stable
